requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
python-slugify>=8.0
